Added simd types to gen/basis-ffi.c and .sml, moved simd files into basis-library/simd and created a prim-simd.sml file
The basis-ffi.c file was a bit weird, all the macros are unhygienic which I suppose is normal for C but threw me off a bit.
I defined the simd types under a #ifdef __x86_64__ which is how I've been defining most simd stuff in c.

I've yet to actually do anything  with the simd stuff in the basis library, but i felt it was a good time to make a commit before I started working on them