Fixed all match errors, everything compiles with no warnings and the compiler runs...well, it runs if not given any input
If given a file the compiler fails with an error that it can't find the simd structures, which makes sense as I've yet to add the simd files in the
basis library to any .mlb file. But things look somewhat promising, I'm not 100% sure about what to do to add the simd structures
but I should be able to figure it out if I poke around in the basis library a bit.

The match errors I was getting before were a bit weird, they were coming from the x86 codegen, it makes sense that there would be match errors as I didn't add the simd c typse to
the x86 codegen until today, but I don't know why anything was being run from there in the first place.