added simd information to the native codegen to allow calling the c-backend from native code
Because I still have an issue with the simd word imports I haven't tested things
yet so I'm not sure I've covered everything, but the compiler proper builds
fine.

For the word issue, I think I might need to go back and make the c types
for simd words be simd128/256_WordN for N=8,16,32,64 instead of just using
simd128/256_WordX, because I think the issue comes down to the c types not
matching the allowed types for import.