worked on load-store primitives a bit & wrote some basic profiling tests
I had a bit of trouble with the timer but profiling things shows a marked
increase in speed for simd operations vs scalar. However I just did a really
basic test to sum up a list of numbers, so I still  need to do more extensive testing
but it is good to see that things are faster.