Got the native codgen working, after some work on array offsets and the like
I realized I made a really stupid misteak, for the scale issue I never needed
to have a scale 16, since the index for loading from an array is based on
the element type of that array, I guess it'll be an issue for making arrays
of simd types, but I think thats ok. As for performance the native code beats
both c and sml for doing a simple loop and a dot product, which it should,
but its nice to see.

I only have real types working for now, so I still need to add word types
and I have a bunch of fine tuning to do, but its nice to see that I've
got things working