added the ability to load simdWords from int arrays, also fixed(kinda) a bug from the last commit that prevented simd programs from compiling
The issue is with trying to create a constant simd value in the basis library.
In this case a constant of all 1s for creating a bitwise not function from the
andnot instruction. The simd-not.sml file in simd-tests shows how it should work
(for 16 bit words, just to demonstrait*) but attempting to do the same thing from
within the basis library causes an option error when attempting to compile any
programs using simd operations. The code for the bitwise not is still there, but
is commented out so that things will work.